+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Check for new Sedge releases on GitHub. Sedge will now report if there is a new version released.
 - Add `--mev-boost-image` flag to specify which Mev Boost docker image sedge uses.
+- Set mev-boost as default on mainnet.
+
+### Changed
+
+- Drop support for deprecated networks. (Kiln, Ropsten, Denver)
+- Update Chiado testnet configs
 
 ### Fixed
 - Print generated files as string instead of []byte